Tweaks for create_scores
